Moved storage hook to main
