+++ components/settings.tsx
@@ -1,32 +1,14 @@
-import { useStorage } from "@plasmohq/storage/hook"
-
-import { categories } from "./data"
 import type { Category } from "./data"
-
-/**
- * Genereate default category weights map for storage
- * @returns Default category weights map
- */
-const defaultCategoryWeightsMap = (): Map<Category, number> => {
-  const map = new Map<Category, number>()
-  categories.forEach((category) => {
-    map.set(category, 1)
-  })
-
-  return map
-}
+import { categories } from "./data"
 
 /**
  * Settings page
  * @returns Settings page
  */
-export const SettingsPage = (): JSX.Element => {
-  // Category storage
-  const [categoryWeights, setCategoryWeights] = useStorage(
-    "gojiCategoryWeights",
-    (stored) => (stored === undefined ? defaultCategoryWeightsMap() : stored)
-  )
-
+export const SettingsPage = (props: {
+  categoryWeights: Map<Category, number>
+  setCategoryWeights: (weights: Map<Category, number>) => void
+}): JSX.Element => {
   return (
     <>
       <h1>Settings</h1>
@@ -41,10 +23,10 @@ export const SettingsPage = (): JSX.Element => {
               type="number"
               min={1}
               max={6}
-              value={categoryWeights[category]}
+              value={props.categoryWeights[category]}
               onChange={(event) =>
-                setCategoryWeights({
-                  ...categoryWeights,
+                props.setCategoryWeights({
+                  ...props.categoryWeights,
                   [category]: event.target.value
                 })
               }
